Implemented mapping and showing forecast data from OpenWeatherMap
Implemented:
- mapping and showing forecast data from OpenWeatherMap

+++ app/src/main/kotlin/ru/mts/avpopo85/weathery/data/model/base/openWeatherMap/forecast/IOWMForecastMainResponse.kt
@@ -22,7 +22,7 @@ interface IOWMForecastMainResponse {
     val tempKf: Double
 
     /**Humidity, %*/
-    val humidity: Double
+    val humidity: Int
 
     /**Atmospheric pressure on the sea level by default, hPa*/
     val atmosphericPressureOnTheSeaLevelByDefaultInhPa: Double
